requirements.txt
openai-whisper
moviepy
openai
ffmpeg-python
dotenv